+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/JSContext.java
@@ -355,6 +355,7 @@ public class JSContext {
         AsyncIteratorFindWithResult,
         AsyncGeneratorAwaitReturnFulfilled,
         AsyncGeneratorAwaitReturnRejected,
+        Crypto,
         IteratorConcat,
         IteratorMap,
         IteratorFilter,