+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The main focus is on user-observable behavior of the engine.
 Changelog may include unreleased versions.
 See [release calendar](https://www.graalvm.org/release-calendar/) for release dates.
 
+## Version 26.0.0
+* Added an experimental option `js.crypto` that provides `getRandomValues()` and `randomUUID()` from the Web Crypto API.
+
 ## Version 25.0.0
 * ECMAScript 2025 mode/features enabled by default.
 * Updated Node.js to version 22.17.1.